+++ src/noComponentPropIntersectionRule.ts
