+++ src/noComponentPropUnionRule.ts
