+++ tests/noComponentPropIntersectionRule.ts
@@ -1,13 +1,13 @@
-import { Rule as NoPropIntersection } from '../src/noPropIntersectionRule';
+import { Rule as NoComponentPropIntersection } from '../src/noComponentPropIntersectionRule';
 import { addFailureSpy, runRuleForTestFile } from './helpers/run-rule-for-test-file';
 
-describe('noPropIntersectionRule', () => {
+describe('noComponentPropIntersectionRule', () => {
   beforeEach(() => {
     addFailureSpy.mockClear();
   });
 
   it('should error once on test file "failure.tsx"', () => {
-    runRuleForTestFile(NoPropIntersection, 'failure.tsx');
+    runRuleForTestFile(NoComponentPropIntersection, 'failure.tsx');
 
     expect(addFailureSpy).toHaveBeenCalledTimes(1);
   });
