+++ tests/noComponentPropUnionRule.ts
@@ -1,13 +1,13 @@
-import { Rule as NoPropUnion } from '../src/noPropUnionRule';
+import { Rule as NoComponentPropUnion } from '../src/noComponentPropUnionRule';
 import { addFailureSpy, runRuleForTestFile } from './helpers/run-rule-for-test-file';
 
-describe('noPropUnionRule', () => {
+describe('noComponentPropUnionRule', () => {
   beforeEach(() => {
     addFailureSpy.mockClear();
   });
 
   it('should error twice on test file "failure.tsx"', () => {
-    runRuleForTestFile(NoPropUnion, 'failure.tsx');
+    runRuleForTestFile(NoComponentPropUnion, 'failure.tsx');
 
     expect(addFailureSpy).toHaveBeenCalledTimes(2);
   });
